fix bug

+++ src/main.py
@@ -95,7 +95,7 @@ def main():
         for (x, y, mask) in tqdm(test_dataloader, '| feature extraction | test | %s |' % class_name):
             test_imgs.extend(x.cpu().detach().numpy())
             gt_list.extend(y.cpu().detach().numpy())
-            gt_mask_list.extend(mask.cpu().detach().numpy())
+            gt_mask_list.extend(mask.cpu().detach().numpy().astype(int))
             # model prediction
             with torch.no_grad():
                 pred = model(x.to(device))
@@ -152,7 +152,7 @@ def main():
             score_map = gaussian_filter(score_map.squeeze().cpu().detach().numpy(), sigma=4)
             score_map_list.append(score_map)
 
-        flatten_gt_mask_list = np.concatenate(gt_mask_list.astype(int)).ravel()
+        flatten_gt_mask_list = np.concatenate(gt_mask_list).ravel()
         flatten_score_map_list = np.concatenate(score_map_list).ravel()
 
         # calculate per-pixel level ROCAUC
